feat: init project
