+++ .editorconfig
@@ -1,0 +1,21 @@
+root = true
+
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*{.ts,.tsx,.js,.jsx}]
+indent_size = 2
+indent_style= space
+trim_trailing_whitespace=true
+insert_final_newline=true
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.yml]
+indent_style = space
+indent_size = 2
