+++ .prettierignore
@@ -1,0 +1,5 @@
+node_modules
+.DS_Store
+dist
+*.local
+node_modules/*
