+++ nodemon.json
@@ -1,0 +1,8 @@
+{
+    "watch": ["src"],
+    "ext": "ts",
+    "ignore": [".git", "node_modules/", "dist/"],
+    "execMap": {
+        "ts": "ts-node-esm"
+    }
+}
